refactor: 잘못된 fixture 패키지 참조 수정
