+++ backend/src/test/java/com/allog/dallog/common/fixtures/OAuthMemberFixtures.java
@@ -6,11 +6,8 @@ public class OAuthMemberFixtures {
 
     public static final String CODE = "authorization code";
 
-    public static final String EMAIL = "example@email.com";
-    public static final String DISPLAY_NAME = "example";
-    public static final String PROFILE_IMAGE_URI = "/image.png";
+    public static final String OAUTH_EMAIL = "example@email.com";
+    public static final String OAUTH_DISPLAY_NAME = "example";
+    public static final String OAUTH_PROFILE_IMAGE_URI = "/image.png";
 
-    public static final String EMAIL2 = "another@email.com";
-    public static final String DISPLAY_NAME2 = "another";
-    public static final String PROFILE_IMAGE_URI2 = "/image.png";
 }
