+++ backend/src/test/java/com/allog/dallog/infrastructure/oauth/client/StubOAuthClient.java
@@ -1,8 +1,8 @@
 package com.allog.dallog.infrastructure.oauth.client;
 
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.DISPLAY_NAME;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.EMAIL;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.PROFILE_IMAGE_URI;
+import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.OAUTH_DISPLAY_NAME;
+import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.OAUTH_EMAIL;
+import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.OAUTH_PROFILE_IMAGE_URI;
 
 import com.allog.dallog.auth.dto.OAuthMember;
 import com.allog.dallog.auth.support.OAuthClient;
@@ -11,6 +11,6 @@ public class StubOAuthClient implements OAuthClient {
 
     @Override
     public OAuthMember getOAuthMember(final String code) {
-        return new OAuthMember(EMAIL, DISPLAY_NAME, PROFILE_IMAGE_URI);
+        return new OAuthMember(OAUTH_EMAIL, OAUTH_DISPLAY_NAME, OAUTH_PROFILE_IMAGE_URI);
     }
 }
