+++ backend/src/test/java/com/allog/dallog/subscription/repository/SubscriptionRepositoryTest.java
@@ -3,12 +3,12 @@ package com.allog.dallog.subscription.repository;
 import static com.allog.dallog.common.fixtures.CategoryFixtures.CATEGORY_1_NAME;
 import static com.allog.dallog.common.fixtures.CategoryFixtures.CATEGORY_2_NAME;
 import static com.allog.dallog.common.fixtures.CategoryFixtures.CATEGORY_3_NAME;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.DISPLAY_NAME;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.DISPLAY_NAME2;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.EMAIL;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.EMAIL2;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.PROFILE_IMAGE_URI;
-import static com.allog.dallog.common.fixtures.OAuthMemberFixtures.PROFILE_IMAGE_URI2;
+import static com.allog.dallog.common.fixtures.MemberFixtures.DISPLAY_NAME;
+import static com.allog.dallog.common.fixtures.MemberFixtures.DISPLAY_NAME2;
+import static com.allog.dallog.common.fixtures.MemberFixtures.EMAIL;
+import static com.allog.dallog.common.fixtures.MemberFixtures.EMAIL2;
+import static com.allog.dallog.common.fixtures.MemberFixtures.PROFILE_IMAGE_URI;
+import static com.allog.dallog.common.fixtures.MemberFixtures.PROFILE_IMAGE_URI2;
 import static com.allog.dallog.common.fixtures.SubscriptionFixtures.COLOR_BLUE;
 import static com.allog.dallog.common.fixtures.SubscriptionFixtures.COLOR_RED;
 import static com.allog.dallog.common.fixtures.SubscriptionFixtures.COLOR_YELLOW;
